Archive 35 Agent/pyproject.toml
[project]
name = "archive-35"
version = "0.1.0"
description = "AI-powered automation for Archive-35 fine art photography business"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "aiogram>=3.13.0",
    "huey>=2.5.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "pillow>=11.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
